+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SoftwareTimer (Timer Lib) inside 14.ADCSensorExample
 - Test for Timer Lib
 - Linear Queue at 19.QueueImplementation
+- Test for linear Queue
 
 ### Changed